[Instrumentor] Implement mechanism to read/write instrumentor database
A instrumentor database file contains the names of the functions and global
variables that have been instrumented for a particular LLVM module. The
database file is named as "<module-name>.instrdb".

The -mllvm -instrumentor-read-database-dir=DIR allows loading the database
files of other instrumented modules that are stored in the directory DIR. The
-mllvm -instrumentor-write-database-dir=DIR allows writing the instrumented
functions and globals from the current module in a database file in the
directory DIR. If these flags are not specified, no database is read or
written.