[LightSan] Add mechanism to support multiple files
Implement adapters as weak symbols to translate pointer arguments. The
sanitized functions define a strong alias with the same name as the weak
adapters. The alias, and thus, the sanitized function will have precedence
over the adapters, and there will be no translation. Otherwise, the adapter
will perform the translation and call the unsanitzed function.